+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.11 (WIP)
 
 * Upgrade to `alchemy-media` v0.6.3
+* Add `purpose` and `mode` attributes to AlchemyForm-based elements for improved field-template getting
 
 ## 0.1.10 (2022-07-14)